- add plugin ressource - add couchdb-river plugin example

+++ manifests/install.pp
@@ -16,6 +16,7 @@ class elasticsearch::install {
   exec { 'download':
     cwd     => '/tmp',
     command => "/usr/bin/wget -q ${elasticsearch::download} -O elasticsearch-${elasticsearch::version}.deb",
+    unless  => "/usr/bin/test -f elasticsearch-${elasticsearch::version}.deb",
     timeout => 120,
   }
